\name{classGraph-package}
\alias{classGraph-package}
\alias{classGraph}
\docType{package}
\title{The R Package 'classGraph'}
\description{
  The package \pkg{classGraph} is package using graph and graph
  visualization methods to visualize inheritance graphs of S4 classes.
}
\details{
%% MM This should be auto-updated from DESCRIPTION !
\tabular{ll}{
Package: \tab classGraph\cr
Type: \tab Package\cr
Version: \tab 0.7-0\cr
Date: \tab 2007-02-10\cr
License: \tab GPL\cr
}
% ~~ An overview of how to use the package, including the most important ~~
% ~~ functions ~~
}
\author{Martin Maechler}
\seealso{
  \code{\link{classTree}()} is the main function of this  package.
}
\keyword{ package }
