\name{classGraph-package}
\alias{classGraph-package}
\alias{classGraph}
\docType{package}
\title{The R Package 'classGraph'}
\description{
  The package \pkg{classGraph} is package using graph and graph
  visualization methods to visualize inheritance graphs of S4 classes.
}
\details{
  %  The DESCRIPTION file:
  \packageDESCRIPTION{classGraph}

  \packageIndices{classGraph}
}
\author{Martin Maechler}
\seealso{
  \code{\link{classTree}()} is the main function of this  package.
}
\keyword{ package }
