\name{superClasses}
\alias{superClasses}
\title{List of Super Classes of a Given S4 Class}
\description{

}
\usage{
superClasses(x)
}
\arguments{
  \item{x}{a class representation as returnd by
    \code{\link[methods]{getClassDef}}.}
}
\value{
  a list of length-1 \code{\link{character}} strings, typically with a
  \code{"package"} attribute each.
}
\author{Robert Gentleman and Martin Maechler}
\seealso{\code{\link{subClasses}}, ...}
\examples{
}
\keyword{classes}
